image-service/requirements.txt
fastapi
uvicorn
pillow
python-multipart
numpy
psutil
onnxruntime
pooch
